src/arm_localizer/requirements.txt
matplotlib==3.4.3
numpy==1.21.2
numpy-base==1.21.2
pillow==8.3.2
python==3.8.11
torch==1.9.0
torchvision=0.10.0

